requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
selenium==4.15.2
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
SQLAlchemy==2.0.23
typing_extensions==4.8.0
greenlet==3.0.1
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
soupsieve==2.5
webdriver-manager==4.0.1 